+++ flutter_project/lib/add_screens/select_route_type_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_project/ptvInfoClasses/route_type_info.dart';
+import 'package:flutter_project/ptv_info_classes/route_type_info.dart';
 import '../ptv_api_service.dart';
 import '../transport.dart';
