+++ flutter_project/lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import "package:flutter_project/addScreens/confirmation_screen.dart";
-import "package:flutter_project/addScreens/select_location_screen.dart";
-import "package:flutter_project/addScreens/select_direction_screen.dart";
-import "package:flutter_project/addScreens/select_route_type_screen.dart";
-import "package:flutter_project/addScreens/select_stop_screen.dart";
+import "package:flutter_project/add_screens/confirmation_screen.dart";
+import "package:flutter_project/add_screens/select_location_screen.dart";
+import "package:flutter_project/add_screens/select_direction_screen.dart";
+import "package:flutter_project/add_screens/select_route_type_screen.dart";
+import "package:flutter_project/add_screens/select_stop_screen.dart";
 // add cupertino for apple version
 
 import "ptv_api_service.dart";
