+++ flutter_project/lib/ptv_info_classes/departure_info.dart
