+++ flutter_project/lib/ptv_info_classes/location_info.dart
