+++ flutter_project/lib/ptv_info_classes/route_info.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_project/ptvInfoClasses/route_direction_info.dart';
+import 'package:flutter_project/ptv_info_classes/route_direction_info.dart';
 
 class Route {
   String id;
