+++ flutter_project/lib/ptv_info_classes/stop_info.dart
